Update to support compact format
The ListingRid didn't get returned in the standard XML format.  This ID field is needed to selecting media for properties.  So I had to implement a Compact parser.  I kept the XML parser, moved it into its own class, and tested incase the ListingRid field isn't important to you.  Side note you can still use Standard XML to save the listingId, different then ListingRid, which can be used to later on select the same house from properties, but it cannot be used for media.