Parse getObject response
Parse the getObject response and return a hash of pictures for the given resource.  The pictures are downloaded to a temporary file and must be  progamically, or manually, deleted after parsing.  I will look into auto deleteing after object is destroyed at a later point but for now my solution will be to auto delete all temp pictures at local midnight.